Do some quick fix on profile page
