+++ app/src/main/java/com/example/caregiver/Identification.java
@@ -48,7 +48,6 @@ public class Identification extends AppCompatActivity {
     /** Navigation function to move to login page **/
     public void openLogIn(View v){
         Intent i = new Intent(Identification.this, Login.class);
-        i.putExtra("tag", tag);
         startActivity(i);
     }
 }
